fix: refresh button also for non-admin user
