+++ src/components/PageRefreshButton/index.ts
@@ -1,0 +1,1 @@
+export * from './PageRefreshButton.tsx';
